requirements.txt
Django==2.0.6
djangorestframework==3.8.2
gunicorn==19.8.1
pytz==2018.4
whitenoise==3.3.1
